Desafio DIO - Ransomware

+++ decrypter.py
@@ -1,22 +1,27 @@
 import os
 import pyaes
 
-## abrir o arquivo criptografado
-file_name = "teste.txt.ransomwaretroll"
-file = open(file_name, "rb")
-file_data = file.read()
-file.close()
-
-## chave para descriptografia
+# chave de descriptografia (deve ser a mesma usada na criptografia)
 key = b"testeransomwares"
 aes = pyaes.AESModeOfOperationCTR(key)
-decrypt_data = aes.decrypt(file_data)
 
-## remover o arquivo criptografado
-os.remove(file_name)
+# listar arquivos no diretório atual
+for file_name in os.listdir():
+    # verificar se o arquivo tem extensão .ransomwaretroll
+    if file_name.endswith(".ransomwaretroll"):
+        # abrir o arquivo criptografado
+        with open(file_name, "rb") as file:
+            crypto_data = file.read()
 
-## criar o arquivo descriptografado
-new_file = "teste.txt"
-new_file = open(f'{new_file}', "wb")
-new_file.write(decrypt_data)
-new_file.close()
+        # descriptografar o conteúdo do arquivo
+        file_data = aes.decrypt(crypto_data)
+
+        # remover o arquivo criptografado
+        os.remove(file_name)
+
+        # restaurar o arquivo original .txt
+        original_file_name = file_name.replace(".ransomwaretroll", "")
+        with open(original_file_name, "wb") as original_file:
+            original_file.write(file_data)
+
+        print(f"Arquivo {original_file_name} restaurado com sucesso!")
